requirements.txt
Django==5.0.6
djangorestframework==3.15.2
mysqlclient==2.2.4
pillow==10.3.0